requirements.txt
torch
torchvision
torchaudio
numpy<2.1
scipy
scikit-learn
timm
pycocotools
faster-coco-eval
transformers
wandb
matplotlib
tensorboard
pyyaml
tqdm
protobuf==3.20.3
huggingface-hub
python-dotenv
lightly-train==0.11.4
ultralytics
